dylib: connect git_revision
