+++ crates/phala-git-revision/Cargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "phala-git-revision"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
