+++ crates/phala-git-revision/src/lib.rs
@@ -1,0 +1,5 @@
+#![no_std]
+
+pub fn git_revision() -> &'static str {
+    env!("PHALA_GIT_REVISION")
+}
