+++ crates/pink/runtime/Cargo.toml
@@ -61,6 +61,7 @@ environmental = "1.1.3"
 phala-sanitized-logger = { path = "../../phala-sanitized-logger" }
 hash-db = "0.15"
 anyhow = "1"
+phala-git-revision = { path = "../../phala-git-revision" }
 
 [dev-dependencies]
 insta = "1.7.2"
